+++ hc/front/templatetags/hc_extras.py
@@ -35,6 +35,18 @@ def site_root():
     return settings.SITE_ROOT
 
 
+@register.simple_tag
+def debug_warning():
+    if settings.DEBUG:
+        return mark_safe("""
+            <div id="debug-warning">
+            Running in debug mode, do not use in production.
+            </div>
+        """)
+
+    return ""
+
+
 def naturalize_int_match(match):
     return '%08d' % (int(match.group(0)),)